Refactor `serialize`
